+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.12.0 - Breaking Change
+* Removed `moduleQualifier` from `FlutterDDIModuleDefine`.
+* `routeQualifier` is now the module qualifier API in `FlutterDDIModuleDefine`.
+
+** Breaking Changes **
+* Removed `FlutterDDIListen` and `.listen` extension.
+
 ## 0.11.1
 * Fixed behavior when modules are destroyed.
 * Fixed a possible memory leak when using interceptors.